+++ LugenStore.API/Services/UserService.cs
@@ -9,7 +9,6 @@ namespace LugenStore.API.Services;
 
 public partial class UserService(IUserRepository _repository) : IUserService
 {
-
     private static void Normalize(UserBaseDto dto)
     {
         dto.Name = dto.Name.Trim();
@@ -64,6 +63,9 @@ public partial class UserService(IUserRepository _repository) : IUserService
 
     public async Task<bool> DeleteAsync(Guid id)
     {
+        if (id == Guid.Empty)
+            throw new ValidationException("Id cannot be empty");
+
         var deleted = await _repository.DeleteAsync(id);
 
         if (!deleted)
